Rename files and folder part 2
