+++ frontend/app/src/pages/overview/OverviewPage.tsx
@@ -6,7 +6,7 @@ import {
     getActuatorsOverviewTest,
     getSensorsOverviewTest,
     SensorsOverview,
-} from "../../shared/services/box-data
+} from "shared/services/box-data";
 
 import {
     Box,
