+++ frontend/app/src/shared/components/Graph/SeriesGraph.tsx
