+++ frontend/app/src/shared/services/box-data.ts
@@ -1,4 +1,4 @@
-import { config } from "../core/config";
+import { config } from "shared/config/api";
 
 export const getGroundHumidity = (
     boxId: string
